refactor: 컴포넌트 및 소스 폴더 구조 도메인별 재편
- components/를 auth/common/home/place/post/user/ 도메인 폴더로 분리
- hooks/에 place/ 서브폴더 추가 (useBookmark, useFacilityDetail 등)
- supabase/query/를 place/와 post/ 서브폴더로 분리
- app/util/time.ts → lib/time.ts 이동
- 미사용 파일 삭제: App.tsx, App.css, index.css, lib/supabase.ts, ui2/sonner.tsx, data/facilities.tsx
- 모든 import 경로 업데이트
