+++ src/app/components/auth/LoginWrapper.tsx
@@ -1,4 +1,4 @@
-import { AuthProvider } from '../contexts/AuthContext'
+import { AuthProvider } from '../../contexts/AuthContext'
 import { Login } from './Login'
 
 export function LoginWrapper() {
