+++ src/app/components/place/FacilityDetail.tsx
@@ -1,12 +1,12 @@
 import { ArrowLeft, Heart, MapPin, Phone } from 'lucide-react'
 import { useState } from 'react'
 import { Link, useParams } from 'react-router'
-import { useBookmark } from '../../hooks/useBookmark'
-import { useFacilityDetail } from '../../hooks/useFacilityDetail'
-import { useFacilityMeta } from '../../hooks/useFacilityMeta'
-import { useReviews } from '../../hooks/useReviews'
-import { useAuth } from '../contexts/AuthContext'
-import { ImageWithFallback } from './figma/ImageWithFallback'
+import { useBookmark } from '../../../hooks/place/useBookmark'
+import { useFacilityDetail } from '../../../hooks/place/useFacilityDetail'
+import { useFacilityMeta } from '../../../hooks/place/useFacilityMeta'
+import { useReviews } from '../../../hooks/place/useReviews'
+import { useAuth } from '../../contexts/AuthContext'
+import { ImageWithFallback } from '../common/ImageWithFallback'
 import { ReviewAverages } from './review/ReviewAverages'
 import { ReviewForm } from './review/ReviewForm'
 import { ReviewList } from './review/ReviewList'
