+++ src/app/components/post/PostDetailPage.tsx
@@ -1,7 +1,7 @@
 import { usePostData } from '@/hooks/queries/use-post-data'
 import { Navigate, useParams } from 'react-router'
-import Fallback from '../fallback'
-import Loader from '../loader'
+import Fallback from '../common/ErrorFallback'
+import Loader from '../common/LoadingSpinner'
 import CommentEditor from './comment/comment-editor'
 import CommentList from './comment/comment-list'
 import PostItem from './post-item'
