--- src/app/lib/supabase.ts
@@ -1,6 +1,0 @@
-import { type Database } from '@/database.types'
-import { createClient } from '@supabase/supabase-js'
-const supabaseUrl = import.meta.env.VITE_SUPABASE_URL
-const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY
-
-export const supabase = createClient<Database>(supabaseUrl, supabaseAnonKey)
