+++ src/app/routes.tsx
@@ -1,15 +1,15 @@
 import { createBrowserRouter } from 'react-router'
-import { Bookmark } from './components/bookmark'
-import { FacilityDetail } from './components/FacilityDetail'
-import { Home } from './components/Home'
-import { LoginWrapper as Login } from './components/LoginWrapper'
-import { MyPage } from './components/MyPage'
-import { NotFound } from './components/NotFound'
-import { Community } from './components/post/Community'
+import { LoginWrapper as Login } from './components/auth/LoginWrapper'
+import { RegisterWrapper as Register } from './components/auth/RegisterWrapper'
+import { Root } from './components/common/Root'
+import { NotFound } from './components/common/NotFound'
+import { Home } from './components/home/Home'
+import { FacilityDetail } from './components/place/FacilityDetail'
+import { PostListPage } from './components/post/PostListPage'
 import PostDetailPage from './components/post/PostDetailPage'
-import { RegisterWrapper as Register } from './components/RegisterWrapper'
-import { ReviewPage } from './components/ReviewPage'
-import { Root } from './components/Root'
+import { MyPage } from './components/user/MyPage'
+import { ReviewPage } from './components/user/ReviewPage'
+import { Bookmark } from './components/user/Bookmark'
 
 export const router = createBrowserRouter([
     {
@@ -19,7 +19,7 @@ export const router = createBrowserRouter([
             { index: true, Component: Home },
             { path: 'facility/:id', Component: FacilityDetail },
             { path: 'mypage', Component: MyPage },
-            { path: 'community', Component: Community },
+            { path: 'community', Component: PostListPage },
             { path: 'post/:id', Component: PostDetailPage },
             { path: 'bookmark', Component: Bookmark },
             { path: 'reviews', Component: ReviewPage },
