+++ src/hooks/mutations/post/comment/use-create-comment.ts
@@ -1,4 +1,4 @@
-import { createComment } from '@/supabase/query/post-comments'
+import { createComment } from '@/supabase/query/post/post-comments'
 import type { UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
