+++ src/hooks/mutations/post/comment/use-delete-comment.ts
@@ -1,4 +1,4 @@
-import { deleteComment } from '@/supabase/query/post-comments'
+import { deleteComment } from '@/supabase/query/post/post-comments'
 import type { UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
