+++ src/hooks/mutations/post/use-create-post.ts
@@ -1,5 +1,5 @@
 import { queryKeys } from '@/hooks/queries/query-keys'
-import { createPost } from '@/supabase/query/post'
+import { createPost } from '@/supabase/query/post/post'
 import type { UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
