+++ src/hooks/mutations/post/use-delete-post.ts
@@ -1,5 +1,5 @@
 import { queryKeys } from '@/hooks/queries/query-keys'
-import { deletePost } from '@/supabase/query/post'
+import { deletePost } from '@/supabase/query/post/post'
 import type { UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
