+++ src/hooks/mutations/post/use-toggle-post-like.ts
@@ -1,5 +1,5 @@
 import { queryKeys } from '@/hooks/queries/query-keys'
-import { togglePostLike } from '@/supabase/query/post'
+import { togglePostLike } from '@/supabase/query/post/post'
 import type { Post, UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
