+++ src/hooks/mutations/post/use-update-post.ts
@@ -1,5 +1,5 @@
 import { queryKeys } from '@/hooks/queries/query-keys'
-import { updatePost } from '@/supabase/query/post'
+import { updatePost } from '@/supabase/query/post/post'
 import type { UseMutationCallback } from '@/types'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
