+++ src/hooks/queries/use-posts-data.ts
@@ -1,4 +1,4 @@
-import { fetchPosts } from '@/supabase/query/post'
+import { fetchPosts } from '@/supabase/query/post/post'
 import { useQuery } from '@tanstack/react-query'
 import { queryKeys } from './query-keys'
